[core][presenters-dispatcher] fix: move dispatcher to lib sources
MoviperPresentersDispatcher
