+++ moviper/src/main/java/com/mateuszkoslacz/moviper/server/MoviperPresentersDispatcher.java
@@ -1,4 +1,4 @@
-package com.mateuszkoslacz.moviper.rxsample.viper.server;
+package com.mateuszkoslacz.moviper.server;
 
 import android.app.Activity;
 import android.app.Fragment;
