+++ sample-super-rx-ai/src/main/java/com/mateuszkoslacz/moviper/rxsample/viper/routing/ListingRouting.java
@@ -6,9 +6,9 @@ import com.mateuszkoslacz.moviper.base.routing.BaseRxRouting;
 import com.mateuszkoslacz.moviper.rxsample.viper.contract.ListingContract;
 import com.mateuszkoslacz.moviper.rxsample.viper.entity.User;
 import com.mateuszkoslacz.moviper.rxsample.viper.presenter.UserDetailsPresenter;
-import com.mateuszkoslacz.moviper.rxsample.viper.server.ActivityStarter;
-import com.mateuszkoslacz.moviper.rxsample.viper.server.MoviperPresentersDispatcher;
 import com.mateuszkoslacz.moviper.rxsample.viper.view.activity.UserDetailsActivity;
+import com.mateuszkoslacz.moviper.server.ActivityStarter;
+import com.mateuszkoslacz.moviper.server.MoviperPresentersDispatcher;
 
 public class ListingRouting
         extends BaseRxRouting
